Clarify a comment about how we hash into a G1 curve point.
The previous comment said we were using the algorithm of Chen and Li,
but that actually chooses the minimum of the two possible y-coords
(we're taking a modular square root, so we have two choices).
In reality, the AMCL function we're using makes the choice based on
least-significant bit.
Maybe these are the same (lsb=0 <=> minimum), but I don't know.
So, just make sure we're clear.