cmake: fix default local dir in find_package scripts
In the Find*.cmake files, the default directory for local builds of
AMCL and xaptum-tpm is specified relative to the ecdaa repo directory
(not the cmake/ subdirectory). So use CMAKE_CURRENT_SOURCE_DIR, not
CMAKE_CURRENT_LIST_DIR.

Also, assume a consistent `build/` subdirectory for the build
directory of the dependencies.  FindAMCL.cmake assumed `install/` and
FindXaptumTPM.cmake assumed `build/`.