test: Remove `trivial_credential_fails` test from signature tests.
This issue came up due to memcheck complaining that non-initialized
memory was used in this test.
It turned out this was due to the AMCL serialization of the point at
infinity quietly failing, and thus not actually setting a BIGNUM when
used as input to a hash that gets converted to the BIGNUM
(the logic between the curve point serialization and the actual
conversion to the BIGNUM is pretty convoluated, so kudos to valgrind for
finding it!).

Not only was this causing memcheck issues, but because the serialization
isn't been done properly this test isn't actually testing anything
(the test is that a verification fails, which it will because the hash
input is wrong, not because we're catching some malicious behavior, as
the test was intended to do).

This kind of validation of malicious input is done, or assumed to be
done throughout this project, in the deserialization logic.
All the deserialization functions do the standard input point validation
(cf. x9.62 spec).
And, sadly, since the serialization doesn't work for infinity, and
because the curves we're using are mostly (all?) prime-order curves and
thus have no small subgroups, I don't see a good way to test this.
We already have some tests that deserializing bad points fails...