test/tpm: Zero-initialize TCTI context.
This is to silence Memcheck warnings about usage of uninitialized memory
usage in the tpm2_tss library.

But why is it using any part of this struct that it didn't initialize??
This appears to be an issue with the tpm2-tss mssim TCTI.
It doesn't arise when running against the device TCTI,
and the tpm2-tss tests themselves use `calloc`, which zero-initializes
the struct so their tests probably just can't find this issue.