amcl-extensions: Test for identity on cofactor!=1.
When hashing to a point on the G1 curve, if we're using a curve with
non-unit cofactor we multiply by the cofactor to get on the prime-order
subgroup.
However, it's possible that, after multiplying by the cofactor, the
prime-order group element we end up with is the identity!

This is a rare occurance for random hash input, but the
hash-to-curve-point function is used to hash the basename, which is
directly specified by the user and thus isn't randomly-generated.

Note: when using a cofactor==1 curve, we don't need to perform the same
(expensive) identity check, because the AMCL function `setx` that we've
called to create the curve point already checks for identity.