test: Randomize message and basename when fuzzing the Schnorr primitive.
Previously, the only input we were randomizing was the basepoint and
private key (and thus public key).

This also adds a comment in the TPM fuzz test, explaining why we don't
randomize the public/private keys or the basepoint
(it's because the TPM already generated the public key using the canonical basepoint).